# packages/js/openapi/model-proxy/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "dbx-tools model proxy",
    "version": "0.9.13"
  },
  "paths": {
...
    "/v1/embeddings": {
      "post": {
        "operationId": "embeddings",
        "requestBody": {
          "description": "Protocol-specific JSON request body.",
          "content": {
            "application/json": {
              "schema": true
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Protocol response returned as JSON or a raw event stream.",
            "content": {
              "application/json": {
                "schema": true
              },
              "text/event-stream": {
                "schema": true
              }
            }
          }
        }
      }
    },
...
    "/v1/responses": {
      "post": {
        "operationId": "responses",
        "requestBody": {
          "description": "Protocol-specific JSON request body.",
          "content": {
            "application/json": {
              "schema": true
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Protocol response returned as JSON or a raw event stream.",
            "content": {
              "application/json": {
                "schema": true
              },
              "text/event-stream": {
                "schema": true
              }
            }
          }
        }
      }
    },
...
  },
...
}